change float for double for energies and temperatures/fields

+++ file_handler.cpp
@@ -70,7 +70,7 @@ bool OpenLROParametersFile(const char* nombrefile, std::ofstream& output_stream,
  * @param output_stream The std::ofstream object to be initialized and opened.
  * @return bool True if the file was successfully opened, false otherwise.
  */
-bool OpenFinalRedFile(const char* nombrefile, float Hache, float TEMPERA, 
+bool OpenFinalRedFile(const char* nombrefile, double Hache, double TEMPERA, 
                     int count, std::ofstream& output_stream) {
     // Construct the full filename: e.g., "DUMP_cu-al-mn_..._200.0_0.txt"
     // Format Hache with a single significant decimal (one digit after the decimal point)
